broke the skills out into actual components to fix an issue with rendering
