+++ src/components/character/form/SkillsForm.js
@@ -1,10 +1,9 @@
 import React, { useState, useEffect } from "react";
-import ApiManager from "../../../modules/ApiManager"
+import SkillsRow from "./SkillsRow"
 
 const SkillsForm = props => {
   const setCharacterSkills = props.setCharacterSkills;
   const characterSkills = props.characterSkills;
-  const skillList = props.skillList;
   // TODO: filter out skills already chosen for future dropdowns
   // const setSkillList = props.setSkillList;
 
@@ -26,44 +25,6 @@ const SkillsForm = props => {
     setCharacterSkills(stateToChange);
   }
 
-  // FIXME: dropdowns get wiped by re-rendering dom, 
-  // though the info is properly saved 
-  const SkillsDropdown = props => {
-    return (
-      <>
-        <select
-          className="skill-selector"
-          id={`${props.x}:${props.y}`}
-          onChange={handleFieldChange}
-        >
-          {skillList.map(skill => (
-            <option key={skill.id} value={skill.id}>
-              {skill.name}
-            </option>
-          ))}
-        </select>
-      </>
-    )
-  } 
-
-  const SkillsRow = props => {
-    // Making a grid with x and y 
-    // x is the row, y is the column 
-    // x is also the Skill's rating #
-    const x = parseInt(props.row)
-
-    // TODO: figure out how to loop this so you only write it once
-    return (
-      <p className="skills-x"> <strong>+{x}</strong>
-       <SkillsDropdown x={x} y="1"/> 
-       <SkillsDropdown x={x} y="2"/> 
-       <SkillsDropdown x={x} y="3"/> 
-       <SkillsDropdown x={x} y="4"/>
-       <SkillsDropdown x={x} y="5"/>
-      </p>
-    )
-  }
-
   useEffect(() => {
 
   }, [])
@@ -75,12 +36,12 @@ const SkillsForm = props => {
         <div>
           {/* Using a classname to determine the "rating" of a particular chosen skill */}
           {/* TODO: figure out how to loop this so you only write it once*/}
-          <SkillsRow row="6"/>
-          <SkillsRow row="5"/>
-          <SkillsRow row="4"/>
-          <SkillsRow row="3"/>
-          <SkillsRow row="2"/>
-          <SkillsRow row="1"/>
+          <SkillsRow row="6" skillList={props.skillList} handleFieldChange={handleFieldChange}/>
+          <SkillsRow row="5" skillList={props.skillList} handleFieldChange={handleFieldChange}/>
+          <SkillsRow row="4" skillList={props.skillList} handleFieldChange={handleFieldChange}/>
+          <SkillsRow row="3" skillList={props.skillList} handleFieldChange={handleFieldChange}/>
+          <SkillsRow row="2" skillList={props.skillList} handleFieldChange={handleFieldChange}/>
+          <SkillsRow row="1" skillList={props.skillList} handleFieldChange={handleFieldChange}/>
         </div>
       </div>
     </>
